--- src/main/kotlin/com/teamsparta/tikitaka/domain/evaluation/service/EvaluationService.kt
@@ -1,5 +1,0 @@
-package com.teamsparta.tikitaka.domain.evaluation.service
-
-interface EvaluationService {
-    fun calculateAndUpdateScores()
-}
