+++ src/main/kotlin/com/teamsparta/tikitaka/domain/evaluation/service/v3/EvaluationServiceImpl.kt
@@ -5,16 +5,19 @@ import com.teamsparta.tikitaka.domain.common.exception.ModelNotFoundException
 import com.teamsparta.tikitaka.domain.evaluation.dto.EvaluationRequest
 import com.teamsparta.tikitaka.domain.evaluation.dto.EvaluationResponse
 import com.teamsparta.tikitaka.domain.evaluation.repository.EvaluationRepository
+import com.teamsparta.tikitaka.domain.team.repository.TeamRepository
 import com.teamsparta.tikitaka.domain.users.repository.UsersRepository
 import com.teamsparta.tikitaka.infra.security.UserPrincipal
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import java.time.LocalDateTime
 
 @Service
 class EvaluationServiceImpl(
     private val evaluationRepository: EvaluationRepository,
     private val usersRepository: UsersRepository,
+    private val teamRepository: TeamRepository,
 ) : EvaluationService {
 
     @Transactional
@@ -38,5 +41,25 @@ class EvaluationServiceImpl(
         evaluation.updateEvaluation(request)
         evaluation.evaluationStatus = true
         return EvaluationResponse.from(evaluation)
+    }
+
+    @Transactional
+    override fun calculateAndUpdateScores() {
+        val startDate = LocalDateTime.now().minusDays(90)
+        val teams = teamRepository.findAll()
+
+        teams.forEach { team ->
+            val evaluations = evaluationRepository.findEvaluationsForTeamFromLast90Days(team.id!!, startDate)
+            val totalMannerScore = evaluations.sumOf { it.mannerScore }
+            val totalSkillScore = evaluations.sumOf { it.skillScore }
+            val totalAttendanceScore = evaluations.sumOf { it.attendanceScore }
+            val evaluationCount = evaluations.size
+
+            team.mannerScore = if (evaluationCount == 0) 0 else totalMannerScore
+            team.tierScore = if (evaluationCount == 0) 0 else totalSkillScore
+            team.attendanceScore = if (evaluationCount == 0) 0 else totalAttendanceScore
+
+            teamRepository.save(team)
+        }
     }
 }
